Fixed ShallowCopy
Issue was that the ShallowCopy call was working properly on the top
level class but not on the Content ArrayList. The ArrayList was
maintaininga reference to the previous ArrayList on ShallowCopy call
which means anytime a ShallowCopy call was made...it'd have a reference
to the old array. This was causing havoc. Now creating a new array
instance and copying the items over to it. This should work as intended
now.