Added some sugar fields which will give the fields in Inches or Pounds (For us US folk) and gave the ability to use larger boxes first
The larger boxes first is the star of the show (for now). It requires
more tinkering and testing...but essentially when packaging items, it'll
attempt to load all the items in to the biggest boxes first....this will
allow less boxes...and perhaps cheaper rates when using a rate lookup
service. But this is the start...need to do more testing.